\name{job.Batch}
\alias{job.Batch}
\docType{data}
\title{
File created by the functions
}

\description{
The first job created by the core function VarEff is a job.Batch
}

\format{
Core of estimation by the model VarEff
  }

\details{
This infile created by the function VarEff is used by the others functions: NatSizeDist, LogSizeDist, NTDist. 
}

\source{
See SUMMARY Manual for VAREFF package: R software from Nikolic and Chevalet.
\url{https://qgp.jouy.inra.fr}
}

\references{
Nikolic N., Chevalet C. 2012. SUMMARY Manual for VAREFF package: R software.
}

\keyword{datasets}
