\name{job.Ndist}
\alias{job.Ndist}
\docType{data}
\title{
File created by the function NatSizeDist
}

\description{
Posterior distribution of effective size (Ne).}

\format{
Big file containing the detailed densities of the posterior distributions of Ne, with 512 points.
  }

\details{
This infile created by the function VarEff is used by the function: plotNdistrib. 
}

\source{
See SUMMARY Manual for VAREFF package: R software from Nikolic and Chevalet.
\url{https://qgp.jouy.inra.fr}
}

\references{
Nikolic N., Chevalet C. 2012. SUMMARY Manual for VAREFF package: R software.
}

\keyword{datasets}
